split packages and update imports
