+++ eth/eventhandler/validation_test.go
@@ -10,8 +10,8 @@ import (
 
 	operatorstorage "github.com/ssvlabs/ssv/operator/storage"
 	"github.com/ssvlabs/ssv/registry/storage"
+	kv "github.com/ssvlabs/ssv/storage/badger"
 	"github.com/ssvlabs/ssv/storage/basedb"
-	"github.com/ssvlabs/ssv/storage/kv"
 )
 
 func Test_validateValidatorAddedEvent(t *testing.T) {
