+++ storage/badger/logger.go
