+++ storage/badger/logger_test.go
